Add redun tests
